+++ influxql/query_executor.go
@@ -2,16 +2,16 @@ package influxql
 
 import (
 	"errors"
-	"expvar"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"log"
 	"runtime/debug"
 	"sync"
+	"sync/atomic"
 	"time"
 
-	"github.com/influxdata/influxdb"
+	"github.com/influxdata/influxdb/models"
 )
 
 var (
@@ -119,7 +119,7 @@ type QueryExecutor struct {
 	Logger *log.Logger
 
 	// expvar-based stats.
-	statMap *expvar.Map
+	stats *QueryStatistics
 }
 
 // NewQueryExecutor returns a new instance of QueryExecutor.
@@ -127,8 +127,26 @@ func NewQueryExecutor() *QueryExecutor {
 	return &QueryExecutor{
 		TaskManager: NewTaskManager(),
 		Logger:      log.New(ioutil.Discard, "[query] ", log.LstdFlags),
-		statMap:     influxdb.NewStatistics("queryExecutor", "queryExecutor", nil),
+		stats:       &QueryStatistics{},
 	}
+}
+
+// QueryStatistics keeps statistics related to the QueryExecutor.
+type QueryStatistics struct {
+	ActiveQueries          int64
+	QueryExecutionDuration int64
+}
+
+// Statistics returns statistics for periodic monitoring.
+func (e *QueryExecutor) Statistics(tags map[string]string) []models.Statistic {
+	return []models.Statistic{{
+		Name: "queryExecutor",
+		Tags: tags,
+		Values: map[string]interface{}{
+			statQueriesActive:          atomic.LoadInt64(&e.stats.ActiveQueries),
+			statQueryExecutionDuration: atomic.LoadInt64(&e.stats.QueryExecutionDuration),
+		},
+	}}
 }
 
 // Close kills all running queries and prevents new queries from being attached.
@@ -154,10 +172,10 @@ func (e *QueryExecutor) executeQuery(query *Query, opt ExecutionOptions, closing
 	defer close(results)
 	defer e.recover(query, results)
 
-	e.statMap.Add(statQueriesActive, 1)
+	atomic.AddInt64(&e.stats.ActiveQueries, 1)
 	defer func(start time.Time) {
-		e.statMap.Add(statQueriesActive, -1)
-		e.statMap.Add(statQueryExecutionDuration, time.Since(start).Nanoseconds())
+		atomic.AddInt64(&e.stats.ActiveQueries, -1)
+		atomic.AddInt64(&e.stats.QueryExecutionDuration, time.Since(start).Nanoseconds())
 	}(time.Now())
 
 	qid, task, err := e.TaskManager.AttachQuery(query, opt.Database, closing)