+++ CHANGELOG.md
@@ -31,6 +31,7 @@ With this release the systemd configuration files for InfluxDB will use the syst
 - [#6889](https://github.com/influxdata/influxdb/pull/6889): Update help and remove unused config options from the configuration file.
 - [#6900](https://github.com/influxdata/influxdb/pull/6900): Trim BOM from Windows Notepad-saved config files.
 - [#6938](https://github.com/influxdata/influxdb/issues/6938): Added favicon
+- [#6507](https://github.com/influxdata/influxdb/issues/6507): Refactor monitor service to avoid expvar and write monitor statistics on a truncated time interval.
 
 ### Bugfixes